Refactor the image files
