+++ src/stories/illustrations.stories.js
@@ -1,9 +1,9 @@
 import React from "react"
 import { Box, Image } from "grommet"
 // importing illustrations
-import image1 from "../images/illustrations/Girl on a chair.svg"
-import image2 from "../images/illustrations/guidance form.svg"
-import image3 from "../images/illustrations/Online classes.svg"
+import image1 from "../../static/illustrations/Girl on a chair.svg"
+import image2 from "../../static/illustrations/guidance form.svg"
+import image3 from "../../static/illustrations/Online classes.svg"
 
 export default {
   title: "Design/Graphics",
